requirements.txt
tensorflow==2.17.0
streamlit==1.37.0
numpy
Pillow
matplotlib
seaborn
scikit-learn